See description.
- Added more KTime converter functions.
- Added NEVER and FOREVER KTime constants.
- Changed KTime#toMillis function to return Double instead of Long.

+++ src/main/kotlin/bvanseg/kotlincommons/lang/thread/ThreadExtensions.kt
@@ -2,4 +2,4 @@ package bvanseg.kotlincommons.lang.thread
 
 import bvanseg.kotlincommons.time.api.KTime
 
-fun Thread.sleep(time: KTime) = Thread.sleep(time.toMillis())
+fun Thread.sleep(time: KTime) = Thread.sleep(time.toMillis().toLong())
